Chunk 10: history, settings and onboarding
Fourteen interface primitives and nothing else. Row and Panel carry most
of the application - a settings row and a transcript row are the same
shape, so they are the same component, and Settings contributes no
components of its own.

Settings is a sticky section list beside one continuous scroll rather
than tabs. Eight groups is too many across the top at this width, and
tabs would hide whichever section contains the thing currently blocking
the user - a missing permission, a model that will not load - behind a
click. Rebinding captures a real key press rather than asking someone to
type Ctrl+Shift+Space into a text box, which is also how they discover
that a chord of two ordinary keys is not something an OS can register.

Onboarding is three steps: get the model, grant the microphone, try it
once. The faster route to understanding a dictation tool is dictating a
sentence, not reading about it.

Two wire-shape corrections came out of building against the real types.
Specta maps every Rust float to a nullable number, because NaN cannot be
represented in JSON - so the f64 byte counts and timestamps chosen
earlier would have pushed a null into every call site. Byte counts are
now u32, ids are u32, and timestamps cross as RFC 3339 strings, which is
exact, self-describing and parsed natively by Date.

Default hotkeys moved off Alt+Space, which is verified bound to
activate-window-menu in Cinnamon on this machine and is the window menu
on Windows too. Ctrl+Shift+Space and Ctrl+Alt+Space are unclaimed on all
three platforms. Bare function keys are now accepted as bindings - they
never appear in prose, so they cannot fire mid-sentence, and one key is
far easier to hold than a three-key chord - and F9/F10 is offered as a
preset.

All three surfaces verified running on this machine.

110 tests, clippy clean, ESLint clean.

+++ src-tauri/src/models/store.rs
@@ -19,14 +19,14 @@ use crate::error::{Error, Result};
 /// one.
 pub const PART_SUFFIX: &str = ".part";
 
-#[derive(Debug, Clone, PartialEq, Serialize, Type)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Type)]
 #[serde(rename_all = "camelCase", tag = "state", content = "detail")]
 pub enum InstallState {
     /// Nothing on disk.
     Missing,
-    /// A download was interrupted; bytes already fetched. `f64` for the same reason
-    /// as `DownloadProgress` — it is what a JavaScript number is.
-    Partial { downloaded: f64 },
+    /// A download was interrupted; bytes already fetched. See `DownloadProgress` for
+    /// why byte counts cross the boundary as `u32`.
+    Partial { downloaded: u32 },
     /// All files present at the expected sizes.
     Installed,
 }
@@ -89,7 +89,7 @@ impl ModelStore {
             InstallState::Missing
         } else {
             InstallState::Partial {
-                downloaded: downloaded as f64,
+                downloaded: downloaded as u32,
             }
         }
     }
@@ -222,7 +222,7 @@ mod tests {
         std::fs::write(dir.join(DEFAULT.files[3].path), b"not the real tokens").unwrap();
 
         match store.state(DEFAULT) {
-            InstallState::Partial { downloaded } => assert!(downloaded > 0.0),
+            InstallState::Partial { downloaded } => assert!(downloaded > 0),
             other => panic!("expected Partial, got {other:?}"),
         }
         let _ = std::fs::remove_dir_all(&root);
